\docType{methods}
\name{fixef}

\alias{fixef}

\title{
  Extract fixed-effects estimates
}

\description{
  Extract the fixed-effects estimates
}

\usage{
  fixef(object, ...)
}

\arguments{
  \item{object}{any fitted model object from which fixed
    effects estimates can be extracted.}
  \item{\dots}{optional additional arguments.}
}

\value{
  a named, numeric vector of fixed-effects estimates.
}

\details{
  Extract the estimates of the fixed-effects parameters
  from a fitted model.
}

\seealso{
  \code{\link{olmm}}, \code{\link{olmm-class}}
}

\keyword{methods}