\docType{methods}
\name{ranef}

\alias{ranef}

\title{
  Extract random-effect estimates
}

\description{
  Extract the random-effects estimates from a fitted model with random
  effects. 
}

\usage{
  ranef(object, ...)
}

\arguments{
  \item{object}{an object of a class of fitted models with
    random effects.}
  \item{\dots}{other arguments to be passed.}
}

\seealso{
  \code{\link{olmm}}, \code{\link{olmm-class}}
}

\keyword{methods}