\docType{methods}
\name{ranefCov}

\alias{ranefCov}

\title{Extract estimated covariance matrix of random effects}

\description{A generic method for computing the estimated variances and
  covariances between the random-effect terms in a linear mixed-effect
  model.} 

\usage{
ranefCov(object, ...)
}

\arguments{
  \item{object}{a fitted model object with random effect components.} 
  \item{...}{other arguments to be passed.}
}

\seealso{
  \code{\link{olmm}}, \code{\link{olmm-class}}
}

\keyword{methods}