\name{neglogLik}

\alias{neglogLik}

\title{Extract the negative log-Likelihood of a model.}

\description{Generic function to extract the log-likelihood of a model
  multiplied by minus one.} 

\usage{
neglogLik(object, ...)
}

\arguments{
  \item{object}{a fitted model object.} 
  \item{...}{other arguments to be passed.}
}

\seealso{\code{\link{logLik}}}