\name{extract}

\alias{extract}

\title{
  Extract characteristics of a fitted object.
}

\description{
  Extract various characteristics of a fitted object.
}

\usage{
extract(object, ...)
}

\arguments{
  \item{object}{a fitted object.} 
  \item{...}{further arguments passed to or from other methods.}
}

\seealso{\code{\link{extract.tvcolmm}}}

\keyword{methods}